Input: adp5589-keys: add support for fw properties
Use firmware properties (eg: OF) to get the device specific
configuration. This change just replaces the platform data since there
was no platform using it and so, it makes no sense having both.

Special note to the GPIO bias and debounce  settings that are now supported
as part of the gpio subsystem (using 'set_config()' callback).

Signed-off-by: Nuno Sa <nuno.sa@analog.com>