Input: adp5589: refactor and platform_data removal
This series aims to remove platform data dependency from the adp5589
driver (as no platform is really using it) and instead add support for
FW properties. Note that rows and columns for the keypad are being given
as masks and that was briefly discussed with Dmitry. For context
on why this is being done as mask [1].

The first couple of patches are fixes that we may want to backport...

[1]: https://lore.kernel.org/linux-input/9db96c99c805e615ba40ca7fd3632174d1e8d11f.camel@gmail.com/

To: Dmitry Torokhov <dmitry.torokhov@gmail.com>
To: Mike Frysinger <vapier@gentoo.org>
To: Dmitry Torokhov <dtor@mail.ru>
To: Rob Herring <robh@kernel.org>
To: Michael Hennerich <michael.hennerich@analog.com>
To: Krzysztof Kozlowski <krzk+dt@kernel.org>
To: Conor Dooley <conor+dt@kernel.org>
Cc: linux-input@vger.kernel.org
Cc: devicetree@vger.kernel.org

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20240930-b4-dev-adp5589-fw-conversion-955b2f42da70",
    "prefixes": []
  }
}